requirements.txt
# Django and DRF
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
drf-yasg==1.21.7

# Database
psycopg2-binary==2.9.9
djongo==1.3.6
pymongo==3.12.3
dj-database-url==2.1.0

# Storage
django-storages==1.14.2
boto3==1.28.64
pillow==10.1.0

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Testing
pytest==7.4.3
pytest-django==4.5.2
coverage==7.3.2

# Development
black==23.10.1
isort==5.12.0
flake8==6.1.0
